pyproject.toml
[project]
name = "wechat-mcp-server"
version = "0.1.0"
description = "MCP server for WeChat Public Articles via mptext.top"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"

[project.scripts]
wechat-mcp = "wechat_mcp.server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wechat_mcp"]
